factored out HttpChatServerActions
